Start splitting up CommonUtil
- Move IP owners computation out to TopologyUtil
- Move BGP topology creation out of CommonUtil
- Move OSPF neighbor initialization out of CommonUtil
- Add a separate VRRP test file
