+++ projects/batfish-common-protocol/src/main/java/org/batfish/common/util/CommonUtil.java
@@ -7,15 +7,9 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSetMultimap;
 import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.ImmutableSortedSet;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Ordering;
 import com.google.common.collect.SetMultimap;
 import com.google.common.collect.Sets;
-import com.google.common.graph.ImmutableValueGraph;
-import com.google.common.graph.MutableValueGraph;
-import com.google.common.graph.Network;
-import com.google.common.graph.ValueGraph;
-import com.google.common.graph.ValueGraphBuilder;
 import com.google.common.hash.Hashing;
 import com.google.errorprone.annotations.MustBeClosed;
 import io.opentracing.contrib.jaxrs2.client.ClientTracingFeature;
@@ -45,7 +39,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.IdentityHashMap;
 import java.util.Iterator;
 import java.util.LinkedList;
@@ -58,14 +51,12 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.SortedMap;
 import java.util.SortedSet;
-import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collector;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -84,46 +75,23 @@ import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.batfish.common.BatfishException;
 import org.batfish.common.BfConsts;
-import org.batfish.common.Pair;
-import org.batfish.common.plugin.ITracerouteEngine;
-import org.batfish.datamodel.AclIpSpace;
-import org.batfish.datamodel.BgpActivePeerConfig;
-import org.batfish.datamodel.BgpPassivePeerConfig;
-import org.batfish.datamodel.BgpPeerConfig;
-import org.batfish.datamodel.BgpPeerConfigId;
-import org.batfish.datamodel.BgpProcess;
-import org.batfish.datamodel.BgpSessionProperties;
-import org.batfish.datamodel.BgpSessionProperties.SessionType;
 import org.batfish.datamodel.Configuration;
-import org.batfish.datamodel.DataPlane;
 import org.batfish.datamodel.Edge;
 import org.batfish.datamodel.EmptyIpSpace;
-import org.batfish.datamodel.Flow;
-import org.batfish.datamodel.FlowDisposition;
-import org.batfish.datamodel.FlowTrace;
 import org.batfish.datamodel.IkeGateway;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.InterfaceAddress;
 import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.IpLink;
-import org.batfish.datamodel.IpProtocol;
 import org.batfish.datamodel.IpSpace;
 import org.batfish.datamodel.IpWildcard;
 import org.batfish.datamodel.IpWildcardIpSpace;
 import org.batfish.datamodel.IpWildcardSetIpSpace;
 import org.batfish.datamodel.IpsecVpn;
-import org.batfish.datamodel.NamedPort;
-import org.batfish.datamodel.NetworkConfigurations;
 import org.batfish.datamodel.Prefix;
-import org.batfish.datamodel.Route;
 import org.batfish.datamodel.SubRange;
 import org.batfish.datamodel.Topology;
 import org.batfish.datamodel.UniverseIpSpace;
-import org.batfish.datamodel.Vrf;
 import org.batfish.datamodel.collections.NodeInterfacePair;
-import org.batfish.datamodel.ospf.OspfArea;
-import org.batfish.datamodel.ospf.OspfNeighbor;
-import org.batfish.datamodel.ospf.OspfProcess;
 import org.glassfish.grizzly.http.server.HttpServer;
 import org.glassfish.grizzly.ssl.SSLContextConfigurator;
 import org.glassfish.grizzly.ssl.SSLEngineConfigurator;
@@ -302,290 +270,6 @@ public class CommonUtil {
       return -1;
     }
     return 0;
-  }
-
-  /**
-   * Compute the {@link Ip}s owned by each interface. hostname -&gt; interface name -&gt; {@link
-   * Ip}s.
-   */
-  public static Map<String, Map<String, Set<Ip>>> computeInterfaceOwnedIps(
-      Map<String, Configuration> configurations, boolean excludeInactive) {
-    return computeInterfaceOwnedIps(
-        computeIpInterfaceOwners(computeNodeInterfaces(configurations), excludeInactive));
-  }
-
-  /**
-   * Invert a mapping from {@link Ip} to owner interfaces (Ip -&gt; hostname -&gt; interface name)
-   * to (hostname -&gt; interface name -&gt; Ip).
-   */
-  public static Map<String, Map<String, Set<Ip>>> computeInterfaceOwnedIps(
-      Map<Ip, Map<String, Set<String>>> ipInterfaceOwners) {
-    Map<String, Map<String, Set<Ip>>> ownedIps = new HashMap<>();
-
-    ipInterfaceOwners.forEach(
-        (ip, owners) ->
-            owners.forEach(
-                (host, ifaces) ->
-                    ifaces.forEach(
-                        iface ->
-                            ownedIps
-                                .computeIfAbsent(host, k -> new HashMap<>())
-                                .computeIfAbsent(iface, k -> new HashSet<>())
-                                .add(ip))));
-
-    // freeze
-    return toImmutableMap(
-        ownedIps,
-        Entry::getKey, /* host */
-        hostEntry ->
-            toImmutableMap(
-                hostEntry.getValue(),
-                Entry::getKey, /* interface */
-                ifaceEntry -> ImmutableSet.copyOf(ifaceEntry.getValue())));
-  }
-
-  /**
-   * Invert a mapping from {@link Ip} to owner interfaces (Ip -&gt; hostname -&gt; interface name)
-   * and convert the set of owned Ips into an IpSpace.
-   */
-  public static Map<String, Map<String, IpSpace>> computeInterfaceOwnedIpSpaces(
-      Map<Ip, Map<String, Set<String>>> ipInterfaceOwners) {
-    return toImmutableMap(
-        computeInterfaceOwnedIps(ipInterfaceOwners),
-        Entry::getKey, /* host */
-        hostEntry ->
-            toImmutableMap(
-                hostEntry.getValue(),
-                Entry::getKey, /* interface */
-                ifaceEntry ->
-                    AclIpSpace.union(
-                        ifaceEntry
-                            .getValue()
-                            .stream()
-                            .map(Ip::toIpSpace)
-                            .collect(Collectors.toList()))));
-  }
-
-  /**
-   * Compute a mapping of IP addresses to a set of hostnames that "own" this IP (e.g., as a network
-   * interface address)
-   *
-   * @param configurations {@link Configurations} keyed by hostname
-   * @param excludeInactive Whether to exclude inactive interfaces
-   * @return A map of {@link Ip}s to a set of hostnames that own this IP
-   */
-  public static Map<Ip, Set<String>> computeIpNodeOwners(
-      Map<String, Configuration> configurations, boolean excludeInactive) {
-    return toImmutableMap(
-        computeIpInterfaceOwners(computeNodeInterfaces(configurations), excludeInactive),
-        Entry::getKey, /* Ip */
-        ipInterfaceOwnersEntry ->
-            /* project away interfaces */
-            ipInterfaceOwnersEntry.getValue().keySet());
-  }
-
-  /**
-   * Compute a mapping from IP address to the interfaces that "own" that IP (e.g., as an network
-   * interface address)
-   *
-   * @param enabledInterfaces A mapping of enabled interfaces hostname -&gt; interface name -&gt;
-   *     {@link Interface}
-   * @param excludeInactive whether to ignore inactive interfaces
-   * @return A map from {@link Ip}s to the {@link Interface}s that own them
-   */
-  public static Map<Ip, Map<String, Set<String>>> computeIpInterfaceOwners(
-      Map<String, Set<Interface>> enabledInterfaces, boolean excludeInactive) {
-    Map<Ip, Map<String, Set<String>>> ipOwners = new HashMap<>();
-    Map<Pair<InterfaceAddress, Integer>, Set<Interface>> vrrpGroups = new HashMap<>();
-    enabledInterfaces.forEach(
-        (hostname, interfaces) ->
-            interfaces.forEach(
-                i -> {
-                  if ((!i.getActive() || i.getBlacklisted()) && excludeInactive) {
-                    return;
-                  }
-                  // collect vrrp info
-                  i.getVrrpGroups()
-                      .forEach(
-                          (groupNum, vrrpGroup) -> {
-                            InterfaceAddress address = vrrpGroup.getVirtualAddress();
-                            if (address == null) {
-                              /*
-                               * Invalid VRRP configuration. The VRRP has no source IP address that
-                               * would be used for VRRP election. This interface could never win the
-                               * election, so is not a candidate.
-                               */
-                              return;
-                            }
-                            Pair<InterfaceAddress, Integer> key = new Pair<>(address, groupNum);
-                            Set<Interface> candidates =
-                                vrrpGroups.computeIfAbsent(
-                                    key, k -> Collections.newSetFromMap(new IdentityHashMap<>()));
-                            candidates.add(i);
-                          });
-                  // collect prefixes
-                  i.getAllAddresses()
-                      .stream()
-                      .map(InterfaceAddress::getIp)
-                      .forEach(
-                          ip ->
-                              ipOwners
-                                  .computeIfAbsent(ip, k -> new HashMap<>())
-                                  .computeIfAbsent(hostname, k -> new HashSet<>())
-                                  .add(i.getName()));
-                }));
-    vrrpGroups.forEach(
-        (p, candidates) -> {
-          InterfaceAddress address = p.getFirst();
-          int groupNum = p.getSecond();
-          /*
-           * Compare priorities first. If tied, break tie based on highest interface IP.
-           */
-          Interface vrrpMaster =
-              Collections.max(
-                  candidates,
-                  Comparator.comparingInt(
-                          (Interface o) -> o.getVrrpGroups().get(groupNum).getPriority())
-                      .thenComparing(o -> o.getAddress().getIp()));
-          ipOwners
-              .computeIfAbsent(address.getIp(), k -> new HashMap<>())
-              .computeIfAbsent(vrrpMaster.getOwner().getHostname(), k -> new HashSet<>())
-              .add(vrrpMaster.getName());
-        });
-
-    // freeze
-    return toImmutableMap(
-        ipOwners,
-        Entry::getKey,
-        ipOwnersEntry ->
-            toImmutableMap(
-                ipOwnersEntry.getValue(),
-                Entry::getKey, // hostname
-                hostIpOwnersEntry -> ImmutableSet.copyOf(hostIpOwnersEntry.getValue())));
-  }
-
-  /**
-   * Compute a mapping of IP addresses to the VRFs that "own" this IP (e.g., as a network interface
-   * address).
-   *
-   * @param excludeInactive whether to ignore inactive interfaces
-   * @param enabledInterfaces A mapping of enabled interfaces hostname -&gt; interface name -&gt;
-   *     {@link Interface}
-   * @return A map of {@link Ip}s to a map of hostnames to vrfs that own the Ip.
-   */
-  public static Map<Ip, Map<String, Set<String>>> computeIpVrfOwners(
-      boolean excludeInactive, Map<String, Set<Interface>> enabledInterfaces) {
-
-    Map<String, Map<String, String>> interfaceVrfs =
-        toImmutableMap(
-            enabledInterfaces,
-            Entry::getKey, /* hostname */
-            nodeInterfaces ->
-                nodeInterfaces
-                    .getValue()
-                    .stream()
-                    .collect(
-                        ImmutableMap.toImmutableMap(Interface::getName, Interface::getVrfName)));
-
-    return toImmutableMap(
-        computeIpInterfaceOwners(enabledInterfaces, excludeInactive),
-        Entry::getKey, /* Ip */
-        ipInterfaceOwnersEntry ->
-            toImmutableMap(
-                ipInterfaceOwnersEntry.getValue(),
-                Entry::getKey, /* Hostname */
-                ipNodeInterfaceOwnersEntry ->
-                    ipNodeInterfaceOwnersEntry
-                        .getValue()
-                        .stream()
-                        .map(interfaceVrfs.get(ipNodeInterfaceOwnersEntry.getKey())::get)
-                        .collect(ImmutableSet.toImmutableSet())));
-  }
-
-  /**
-   * Aggregate a mapping (Ip -&gt; host name -&gt; interface name) to (Ip -&gt; host name -&gt; vrf
-   * name)
-   */
-  public static Map<Ip, Map<String, Set<String>>> computeIpVrfOwners(
-      Map<Ip, Map<String, Set<String>>> ipInterfaceOwners, Map<String, Configuration> configs) {
-    return toImmutableMap(
-        ipInterfaceOwners,
-        Entry::getKey, /* ip */
-        ipEntry ->
-            toImmutableMap(
-                ipEntry.getValue(),
-                Entry::getKey, /* node */
-                nodeEntry ->
-                    ImmutableSet.copyOf(
-                        nodeEntry
-                            .getValue()
-                            .stream()
-                            .map(
-                                iface ->
-                                    configs
-                                        .get(nodeEntry.getKey())
-                                        .getAllInterfaces()
-                                        .get(iface)
-                                        .getVrfName())
-                            .collect(Collectors.toList()))));
-  }
-
-  /**
-   * Invert a mapping from Ip to VRF owners (Ip -&gt; host name -&gt; VRF name) and combine all IPs
-   * owned by each VRF into an IpSpace.
-   */
-  public static Map<String, Map<String, IpSpace>> computeVrfOwnedIpSpaces(
-      Map<Ip, Map<String, Set<String>>> ipVrfOwners) {
-    Map<String, Map<String, AclIpSpace.Builder>> builders = new HashMap<>();
-    ipVrfOwners.forEach(
-        (ip, ipNodeVrfs) ->
-            ipNodeVrfs.forEach(
-                (node, vrfs) ->
-                    vrfs.forEach(
-                        vrf ->
-                            builders
-                                .computeIfAbsent(node, k -> new HashMap<>())
-                                .computeIfAbsent(vrf, k -> AclIpSpace.builder())
-                                .thenPermitting(ip.toIpSpace()))));
-
-    return toImmutableMap(
-        builders,
-        Entry::getKey, /* node */
-        nodeEntry ->
-            toImmutableMap(
-                nodeEntry.getValue(),
-                Entry::getKey, /* vrf */
-                vrfEntry -> vrfEntry.getValue().build()));
-  }
-
-  public static Map<Ip, String> computeIpOwnersSimple(Map<Ip, Set<String>> ipOwners) {
-    Map<Ip, String> ipOwnersSimple = new HashMap<>();
-    ipOwners.forEach(
-        (ip, owners) -> {
-          String hostname =
-              owners.size() == 1 ? owners.iterator().next() : Route.AMBIGUOUS_NEXT_HOP;
-          ipOwnersSimple.put(ip, hostname);
-        });
-    return ipOwnersSimple;
-  }
-
-  public static Map<Ip, String> computeIpOwnersSimple(
-      Map<String, Configuration> configurations, boolean excludeInactive) {
-    return computeIpOwnersSimple(computeIpNodeOwners(configurations, excludeInactive));
-  }
-
-  /**
-   * Compute the interfaces of each node.
-   *
-   * @param configurations The {@link Configuration}s for the network
-   * @return A map from hostname to the interfaces of that node.
-   */
-  public static Map<String, Set<Interface>> computeNodeInterfaces(
-      Map<String, Configuration> configurations) {
-    return toImmutableMap(
-        configurations,
-        Entry::getKey,
-        e -> ImmutableSet.copyOf(e.getValue().getAllInterfaces().values()));
   }
 
   public static void copy(Path srcPath, Path dstPath) {
@@ -907,291 +591,6 @@ public class CommonUtil {
     return time;
   }
 
-  /**
-   * Check if a bgp peer is reachable to establish a session
-   *
-   * <p><b>Warning:</b> Notion of directionality is important here, we are assuming {@code src} is
-   * initiating the connection according to its local configuration
-   */
-  private static boolean isReachableBgpNeighbor(
-      BgpPeerConfigId initiator,
-      BgpPeerConfigId listener,
-      BgpActivePeerConfig src,
-      @Nullable ITracerouteEngine tracerouteEngine,
-      @Nullable DataPlane dp) {
-
-    Ip srcAddress = src.getLocalIp();
-    Ip dstAddress = src.getPeerAddress();
-    if (dstAddress == null) {
-      return false;
-    }
-    if (tracerouteEngine == null || dp == null) {
-      throw new BatfishException("Cannot compute neighbor reachability without a dataplane");
-    }
-
-    /*
-     * Ensure that the session can be established by running traceroute in both directions
-     */
-    Flow.Builder fb = new Flow.Builder();
-    fb.setIpProtocol(IpProtocol.TCP);
-    fb.setTag("neighbor-resolution");
-
-    fb.setIngressNode(initiator.getHostname());
-    fb.setIngressVrf(initiator.getVrfName());
-    fb.setSrcIp(srcAddress);
-    fb.setDstIp(dstAddress);
-    fb.setSrcPort(NamedPort.EPHEMERAL_LOWEST.number());
-    fb.setDstPort(NamedPort.BGP.number());
-    Flow forwardFlow = fb.build();
-
-    // Execute the "initiate connection" traceroute
-    SortedMap<Flow, Set<FlowTrace>> traces =
-        tracerouteEngine.processFlows(dp, ImmutableSet.of(forwardFlow), dp.getFibs(), false);
-
-    SortedSet<FlowTrace> acceptedFlows =
-        traces
-            .get(forwardFlow)
-            .stream()
-            .filter(
-                trace ->
-                    trace.getDisposition() == FlowDisposition.ACCEPTED
-                        && trace.getAcceptingNode() != null
-                        && trace.getAcceptingNode().getHostname().equals(listener.getHostname()))
-            .collect(ImmutableSortedSet.toImmutableSortedSet(FlowTrace::compareTo));
-
-    if (acceptedFlows.isEmpty()) {
-      return false;
-    }
-    NodeInterfacePair acceptPoint = acceptedFlows.first().getAcceptingNode();
-    if (SessionType.isEbgp(BgpSessionProperties.getSessionType(src))
-        && !src.getEbgpMultihop()
-        && acceptedFlows.first().getHops().size() > 1) {
-      // eBGP expects direct connection (single hop) unless explicitly configured multi-hop
-      return false;
-    }
-
-    if (acceptPoint == null) {
-      return false;
-    }
-    String acceptedHostname = acceptPoint.getHostname();
-    // The reply traceroute
-    fb.setIngressNode(acceptedHostname);
-    fb.setIngressVrf(listener.getVrfName());
-    fb.setSrcIp(forwardFlow.getDstIp());
-    fb.setDstIp(forwardFlow.getSrcIp());
-    fb.setSrcPort(forwardFlow.getDstPort());
-    fb.setDstPort(forwardFlow.getSrcPort());
-    Flow backwardFlow = fb.build();
-    traces = tracerouteEngine.processFlows(dp, ImmutableSet.of(backwardFlow), dp.getFibs(), false);
-
-    /*
-     * If backward traceroutes fail, do not consider the neighbor reachable
-     */
-    return !traces
-        .get(backwardFlow)
-        .stream()
-        .filter(
-            trace ->
-                trace.getDisposition() == FlowDisposition.ACCEPTED
-                    && trace.getAcceptingNode() != null
-                    && trace.getAcceptingNode().getHostname().equals(initiator.getHostname()))
-        .collect(ImmutableSet.toImmutableSet())
-        .isEmpty();
-  }
-
-  /**
-   * Compute the BGP topology -- a network of {@link BgpPeerConfig}s connected by {@link
-   * BgpSessionProperties}s. See {@link #initBgpTopology(Map, Map, boolean, boolean,
-   * ITracerouteEngine, DataPlane)} for more details.
-   *
-   * @param configurations configuration keyed by hostname
-   * @param ipOwners Ip owners (see {@link #computeIpNodeOwners(Map, boolean)}
-   * @param keepInvalid whether to keep improperly configured neighbors. If performing configuration
-   *     checks, you probably want this set to {@code true}, otherwise (e.g., computing dataplane)
-   *     you want this to be {@code false}.
-   * @return A graph ({@link Network}) representing all BGP peerings.
-   */
-  public static ValueGraph<BgpPeerConfigId, BgpSessionProperties> initBgpTopology(
-      Map<String, Configuration> configurations,
-      Map<Ip, Set<String>> ipOwners,
-      boolean keepInvalid) {
-    return initBgpTopology(configurations, ipOwners, keepInvalid, false, null, null);
-  }
-
-  /**
-   * Compute the BGP topology -- a network of {@link BgpPeerConfigId}s connected by {@link
-   * BgpSessionProperties}s.
-   *
-   * @param configurations node configurations, keyed by hostname
-   * @param ipOwners network Ip owners (see {@link #computeIpNodeOwners(Map, boolean)} for
-   *     reference)
-   * @param keepInvalid whether to keep improperly configured neighbors. If performing configuration
-   *     checks, you probably want this set to {@code true}, otherwise (e.g., computing dataplane)
-   *     you want this to be {@code false}.
-   * @param checkReachability whether to perform dataplane-level checks to ensure that neighbors are
-   *     reachable and sessions can be established correctly. <b>Note:</b> this is different from
-   *     {@code keepInvalid=false}, which only does filters invalid neighbors at the control-plane
-   *     level
-   * @param tracerouteEngine an instance of {@link ITracerouteEngine} for doing reachability checks.
-   * @param dp (partially) computed dataplane.
-   * @return A graph ({@link Network}) representing all BGP peerings.
-   */
-  public static ValueGraph<BgpPeerConfigId, BgpSessionProperties> initBgpTopology(
-      Map<String, Configuration> configurations,
-      Map<Ip, Set<String>> ipOwners,
-      boolean keepInvalid,
-      boolean checkReachability,
-      @Nullable ITracerouteEngine tracerouteEngine,
-      @Nullable DataPlane dp) {
-
-    // TODO: handle duplicate ips on different vrfs
-
-    NetworkConfigurations networkConfigurations = NetworkConfigurations.of(configurations);
-    /*
-     * First pass: identify all addresses "owned" by BgpNeighbors,
-     * add neighbor ids as vertices to the graph
-     */
-    Map<Ip, Set<BgpPeerConfigId>> localAddresses = new HashMap<>();
-    MutableValueGraph<BgpPeerConfigId, BgpSessionProperties> graph =
-        ValueGraphBuilder.directed().allowsSelfLoops(false).build();
-    for (Configuration node : configurations.values()) {
-      String hostname = node.getHostname();
-      for (Vrf vrf : node.getVrfs().values()) {
-        BgpProcess proc = vrf.getBgpProcess();
-        if (proc == null) {
-          // nothing to do if no bgp process on this VRF
-          continue;
-        }
-        for (Entry<Prefix, ? extends BgpPeerConfig> entry :
-            Iterables.concat(
-                proc.getActiveNeighbors().entrySet(), proc.getPassiveNeighbors().entrySet())) {
-          Prefix prefix = entry.getKey();
-          BgpPeerConfig bgpPeerConfig = entry.getValue();
-
-          if (!bgpConfigPassesSanityChecks(bgpPeerConfig, hostname, ipOwners) && !keepInvalid) {
-            continue;
-          }
-
-          BgpPeerConfigId neighborID =
-              new BgpPeerConfigId(
-                  hostname, vrf.getName(), prefix, bgpPeerConfig instanceof BgpPassivePeerConfig);
-          graph.addNode(neighborID);
-
-          // Add this neighbor as owner of its local address
-          localAddresses
-              .computeIfAbsent(bgpPeerConfig.getLocalIp(), k -> new HashSet<>())
-              .add(neighborID);
-        }
-      }
-    }
-
-    // Second pass: add edges to the graph. Note, these are directed edges.
-    for (BgpPeerConfigId neighborId : graph.nodes()) {
-      if (neighborId.isDynamic()) {
-        // Passive end of the peering cannot initiate a connection
-        continue;
-      }
-      BgpActivePeerConfig neighbor = networkConfigurations.getBgpPointToPointPeerConfig(neighborId);
-      if (neighbor == null || neighbor.getPeerAddress() == null) {
-        continue;
-      }
-      Set<BgpPeerConfigId> candidates = localAddresses.get(neighbor.getPeerAddress());
-      if (candidates == null) {
-        // Check maybe it's trying to reach a dynamic neighbor
-        candidates = localAddresses.get(Ip.AUTO);
-        if (candidates == null) {
-          continue;
-        }
-        candidates =
-            candidates
-                .stream()
-                .filter(c -> c.getRemotePeerPrefix().containsIp(neighbor.getPeerAddress()))
-                .collect(ImmutableSet.toImmutableSet());
-        if (candidates.isEmpty()) {
-          // No remote connection candidates
-          continue;
-        }
-      }
-      Long localLocalAs = neighbor.getLocalAs();
-      Long localRemoteAs = neighbor.getRemoteAs();
-      if (localLocalAs == null || localRemoteAs == null) {
-        // AS numbers not configured properly, cannot establish edge.
-        continue;
-      }
-      for (BgpPeerConfigId candidateNeighborId : candidates) {
-        if (!bgpCandidatePassesSanityChecks(neighbor, candidateNeighborId, networkConfigurations)) {
-          // Short-circuit if there is no way the remote end will accept our connection
-          continue;
-        }
-        /*
-         * Perform reachability checks.
-         */
-        if (checkReachability) {
-          if (isReachableBgpNeighbor(
-              neighborId, candidateNeighborId, neighbor, tracerouteEngine, dp)) {
-            graph.putEdgeValue(
-                neighborId,
-                candidateNeighborId,
-                BgpSessionProperties.from(
-                    neighbor,
-                    Objects.requireNonNull(
-                        networkConfigurations.getBgpPeerConfig(candidateNeighborId))));
-          }
-        } else {
-          graph.putEdgeValue(
-              neighborId,
-              candidateNeighborId,
-              BgpSessionProperties.from(
-                  neighbor,
-                  Objects.requireNonNull(
-                      networkConfigurations.getBgpPeerConfig(candidateNeighborId))));
-        }
-      }
-    }
-    return ImmutableValueGraph.copyOf(graph);
-  }
-
-  static boolean bgpConfigPassesSanityChecks(
-      BgpPeerConfig config, String hostname, Map<Ip, Set<String>> ipOwners) {
-    /*
-     * Do these checks as a short-circuit to avoid extra reachability checks when building
-     * BGP topology.
-     * Only keep invalid neighbors that don't have local IPs if specifically requested to.
-     * Note: we use Ip.AUTO to denote the listening end of a dynamic peering.
-     */
-    Ip localAddress = config.getLocalIp();
-    return (localAddress != null
-            && ipOwners.containsKey(localAddress)
-            && ipOwners.get(localAddress).contains(hostname))
-        || Ip.AUTO.equals(localAddress);
-  }
-
-  /**
-   * Check if the given combo of BGP peer configs can agree on their respective BGP local/remote AS
-   * number configurations.
-   */
-  @VisibleForTesting
-  static boolean bgpCandidatePassesSanityChecks(
-      @Nonnull BgpActivePeerConfig neighbor,
-      @Nonnull BgpPeerConfigId candidateId,
-      @Nonnull NetworkConfigurations nc) {
-    if (candidateId.isDynamic()) {
-      BgpPassivePeerConfig candidate = nc.getBgpDynamicPeerConfig(candidateId);
-      return candidate != null
-          && neighbor.getLocalIp() != null
-          && candidate.canConnect(neighbor.getLocalAs())
-          && Objects.equals(neighbor.getRemoteAs(), candidate.getLocalAs())
-          && candidate.canConnect(neighbor.getLocalIp());
-    } else {
-      BgpActivePeerConfig candidate = nc.getBgpPointToPointPeerConfig(candidateId);
-      return candidate != null
-          && Objects.equals(neighbor.getPeerAddress(), candidate.getLocalIp())
-          && Objects.equals(neighbor.getLocalIp(), candidate.getPeerAddress())
-          && Objects.equals(neighbor.getRemoteAs(), candidate.getLocalAs())
-          && Objects.equals(neighbor.getLocalAs(), candidate.getRemoteAs());
-    }
-  }
-
   @VisibleForTesting
   static SetMultimap<Ip, IpWildcardSetIpSpace> initPrivateIpsByPublicIp(
       Map<String, Configuration> configurations) {
@@ -1238,97 +637,6 @@ public class CommonUtil {
               });
     }
     return builder.build();
-  }
-
-  public static void initRemoteOspfNeighbors(
-      Map<String, Configuration> configurations, Map<Ip, Set<String>> ipOwners, Topology topology) {
-    for (Entry<String, Configuration> e : configurations.entrySet()) {
-      String hostname = e.getKey();
-      Configuration c = e.getValue();
-      for (Entry<String, Vrf> e2 : c.getVrfs().entrySet()) {
-        Vrf vrf = e2.getValue();
-        OspfProcess proc = vrf.getOspfProcess();
-        if (proc != null) {
-          proc.setOspfNeighbors(new TreeMap<>());
-          String vrfName = e2.getKey();
-          for (Entry<Long, OspfArea> e3 : proc.getAreas().entrySet()) {
-            long areaNum = e3.getKey();
-            OspfArea area = e3.getValue();
-            for (String ifaceName : area.getInterfaces()) {
-              Interface iface = c.getAllInterfaces().get(ifaceName);
-              if (iface.getOspfPassive()) {
-                continue;
-              }
-              SortedSet<Edge> ifaceEdges =
-                  topology.getInterfaceEdges().get(new NodeInterfacePair(hostname, ifaceName));
-              boolean hasNeighbor = false;
-              Ip localIp = iface.getAddress().getIp();
-              if (ifaceEdges != null) {
-                for (Edge edge : ifaceEdges) {
-                  if (edge.getNode1().equals(hostname)) {
-                    String remoteHostname = edge.getNode2();
-                    String remoteIfaceName = edge.getInt2();
-                    Configuration remoteNode = configurations.get(remoteHostname);
-                    Interface remoteIface = remoteNode.getAllInterfaces().get(remoteIfaceName);
-                    if (remoteIface.getOspfPassive()) {
-                      continue;
-                    }
-                    Vrf remoteVrf = remoteIface.getVrf();
-                    String remoteVrfName = remoteVrf.getName();
-                    OspfProcess remoteProc = remoteVrf.getOspfProcess();
-                    if (remoteProc != null) {
-                      if (remoteProc.getOspfNeighbors() == null) {
-                        remoteProc.setOspfNeighbors(new TreeMap<>());
-                      }
-                      OspfArea remoteArea = remoteProc.getAreas().get(areaNum);
-                      if (remoteArea != null
-                          && remoteArea.getInterfaces().contains(remoteIfaceName)) {
-                        Ip remoteIp = remoteIface.getAddress().getIp();
-                        IpLink localKey = new IpLink(localIp, remoteIp);
-                        OspfNeighbor neighbor = proc.getOspfNeighbors().get(localKey);
-                        if (neighbor == null) {
-                          hasNeighbor = true;
-
-                          // initialize local neighbor
-                          neighbor = new OspfNeighbor(localKey);
-                          neighbor.setArea(areaNum);
-                          neighbor.setVrf(vrfName);
-                          neighbor.setOwner(c);
-                          neighbor.setInterface(iface);
-                          proc.getOspfNeighbors().put(localKey, neighbor);
-
-                          // initialize remote neighbor
-                          IpLink remoteKey = new IpLink(remoteIp, localIp);
-                          OspfNeighbor remoteNeighbor = new OspfNeighbor(remoteKey);
-                          remoteNeighbor.setArea(areaNum);
-                          remoteNeighbor.setVrf(remoteVrfName);
-                          remoteNeighbor.setOwner(remoteNode);
-                          remoteNeighbor.setInterface(remoteIface);
-                          remoteProc.getOspfNeighbors().put(remoteKey, remoteNeighbor);
-
-                          // link neighbors
-                          neighbor.setRemoteOspfNeighbor(remoteNeighbor);
-                          remoteNeighbor.setRemoteOspfNeighbor(neighbor);
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-              if (!hasNeighbor) {
-                IpLink key = new IpLink(localIp, Ip.ZERO);
-                OspfNeighbor neighbor = new OspfNeighbor(key);
-                neighbor.setArea(areaNum);
-                neighbor.setVrf(vrfName);
-                neighbor.setOwner(c);
-                neighbor.setInterface(iface);
-                proc.getOspfNeighbors().put(key, neighbor);
-              }
-            }
-          }
-        }
-      }
-    }
   }
 
   public static void initRemoteIpsecVpns(Map<String, Configuration> configurations) {
