+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/OspfNeighborTest.java
@@ -9,10 +9,10 @@ import static org.hamcrest.Matchers.not;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
-import java.util.Set;
 import org.batfish.common.util.CommonUtil;
 import org.batfish.datamodel.ospf.OspfArea;
 import org.batfish.datamodel.ospf.OspfProcess;
+import org.batfish.datamodel.ospf.OspfTopologyUtils;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -68,9 +68,8 @@ public class OspfNeighborTest {
 
     Map<String, Configuration> configurations =
         ImmutableMap.of(c1.getHostname(), c1, c2.getHostname(), c2);
-    Map<Ip, Set<String>> ipOwners = CommonUtil.computeIpNodeOwners(configurations, true);
     Topology topology = CommonUtil.synthesizeTopology(configurations);
-    CommonUtil.initRemoteOspfNeighbors(configurations, ipOwners, topology);
+    OspfTopologyUtils.initRemoteOspfNeighbors(configurations, topology);
 
     assertThat(op1, hasOspfNeighbors(hasKey(new IpLink(ip1, Ip.ZERO))));
     assertThat(op1, hasOspfNeighbors(not(hasKey(hasIp1(equalTo(ip2))))));
