+++ projects/batfish/src/main/java/org/batfish/grammar/routing_table/nxos/NxosRoutingTableExtractor.java
@@ -9,7 +9,7 @@ import org.antlr.v4.runtime.tree.ParseTreeWalker;
 import org.batfish.common.BatfishException;
 import org.batfish.common.Warnings;
 import org.batfish.common.plugin.IBatfish;
-import org.batfish.common.util.CommonUtil;
+import org.batfish.common.topology.TopologyUtil;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.Ip;
@@ -63,7 +63,7 @@ public class NxosRoutingTableExtractor extends NxosRoutingTableParserBaseListene
     _parser = parser;
     _w = w;
     Map<String, Configuration> configurations = batfish.loadConfigurations();
-    Map<Ip, String> ipOwnersSimple = CommonUtil.computeIpOwnersSimple(configurations, true);
+    Map<Ip, String> ipOwnersSimple = TopologyUtil.computeIpOwnersSimple(configurations, true);
     _ipOwners = ipOwnersSimple;
   }
