+++ projects/question/src/main/java/org/batfish/question/ipowners/IpOwnersAnswerer.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.Set;
 import org.batfish.common.Answerer;
 import org.batfish.common.plugin.IBatfish;
-import org.batfish.common.util.CommonUtil;
+import org.batfish.common.topology.TopologyUtil;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.Ip;
@@ -39,8 +39,8 @@ class IpOwnersAnswerer extends Answerer {
   public AnswerElement answer() {
     IpOwnersQuestion question = (IpOwnersQuestion) _question;
     Map<String, Configuration> configurations = _batfish.loadConfigurations();
-    Map<Ip, Set<String>> ipNodeOwners = CommonUtil.computeIpNodeOwners(configurations, false);
-    Map<String, Set<Interface>> interfaces = CommonUtil.computeNodeInterfaces(configurations);
+    Map<Ip, Set<String>> ipNodeOwners = TopologyUtil.computeIpNodeOwners(configurations, false);
+    Map<String, Set<Interface>> interfaces = TopologyUtil.computeNodeInterfaces(configurations);
 
     TableAnswerElement answerElement = new TableAnswerElement(getTableMetadata());
